requirements.txt
discord.py~=2.0.0
SQLAlchemy~=1.4.13
SQLAlchemy-Utils~=0.37.2
psycopg2-binary~=2.8.6
cryptography~=3.4.6
pytz~=2021.1
matplotlib~=3.3.4
numpy~=1.20.1
requests~=2.25.1
scikit-image~=0.18.1
beautifulsoup4~=4.9.3
lxml~=4.6.2
Markdown~=3.3.4
pytest~=6.2.2
alembic~=1.6.0
PyYAML~=5.4.1
black==22.3.0
isort==5.8.0
requests~=2.25.1
parsita~=1.4.1
pretend~=1.0.9
dateparser~=1.1.1
humanize~=3.5.0
more_itertools~=8.8.0
Pillow~=9.0.1
discord_simple_pretty_help==0.0.1
icalendar~=4.1.0